+++ src/main/java/roomescape/config/ClientConfig.java
@@ -5,6 +5,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestClient;
 import roomescape.exception.PaymentErrorHandler;
 import roomescape.infra.PaymentRestClient;
+import roomescape.infra.PaymentRestClientImpl;
 
 @Configuration
 public class ClientConfig {
@@ -16,7 +17,7 @@ public class ClientConfig {
 
     @Bean
     public PaymentRestClient paymentRestClient() {
-        return new PaymentRestClient(
+        return new PaymentRestClientImpl(
                 RestClient.builder()
                         .baseUrl("https://api.tosspayments.com/v1/payments/confirm")
                         .build(),
